+++ spec/functional/mongoid/config_spec.rb
@@ -36,7 +36,7 @@ describe Mongoid::Config do
   after(:all) do
     Mongoid.configure do |config|
       name          = "mongoid_test"
-      config.master = Mongo::Connection.new(HOST, PORT).db(name)
+      config.master = Mongo::MongoClient.new(HOST, PORT).db(name)
       config.logger = nil
     end
   end
@@ -275,7 +275,7 @@ describe Mongoid::Config do
     context "when provided a mongo database" do
 
       before do
-        described_class.master = Mongo::Connection.new(HOST, PORT).db("mongoid_test")
+        described_class.master = Mongo::MongoClient.new(HOST, PORT).db("mongoid_test")
       end
 
       it "sets the master" do